new: (early) C++ attempt
Tedious, and doesn't seem to be great for organization.

Early progress already feels like an spaghetti monster, because
namespaces and classes are very limited for this task compared to Zig.
Zig's interoperability with C feels more ergonomic.

Also, the RAII of classes gets in the way more than it helps on this
engine since everything is manually managed.